QDoc: Add INTERNAL_MODULE to exported test modules
QDoc recently moved certain test-libraries that it uses under src.
To satisfy the CI requirements and allow the usages of those libraries
during testing, the libraries were exported through
`qt_internal_add_module`.

The libraries are intended for use only in QDoc, and generally only
during testing, and are not intended for outside usage.

Hence, mark them with `INTERNAL_MODULE` in their interface target
definition.

Due to the addition of `INTERNAL_MODULE`, the final include path for those
targets will change when they are installed, losing the "Private" suffix
in their name.

Hence, the `PUBLIC_INCLUDE_DIRECTORIES` for the installation was
modified to refer to the new path.

Change-Id: If63524aad2923dab628c85e20f59f698c3f38ab8
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
Reviewed-by: Paul Wicking <paul.wicking@qt.io>