qdoc: Set correct genus for TypeAliasNode instances
QDoc is meant to treat nodes constructed via \typedef and \typealias
commands identically; however, some switch-statements in the code
omitted a case for Node::TypeAlias, leading to incorrect behavior.

Specifically, marked-up code did nothing for a type alias, and the
'genus' for a TypeAliasNode defaulted to 'DontCare'. This meant that even though normal linking to type alias documentation works,
autolinking of the marked-up code in function signatures
failed for type aliases as there, QDoc requires the linked node
to match the genus of the function (in this case, 'CPP').

Pick-to: 6.6 6.5
Fixes: QTBUG-113863
Change-Id: I9db6727b5e91c30cad14d233eff0319a5dde73c1
Reviewed-by: Luca Di Sera <luca.disera@qt.io>
Reviewed-by: Paul Wicking <paul.wicking@qt.io>