CMake: Don't delete generated .ts files on "ninja clean"
When adding a translation, the corresponding .ts file doesn't exist yet.
Still, we need the .ts file to satisfy the *_lrelease targets we're
running on every build.

We had a custom command that was generating an initial .ts file in the
source directory with lupdate. This had the downside that "ninja clean"
would remove this source file - even after the user modified it. On the
upside, lupdate was able to guess the language from the file name and
wrote the language attribute to the generated .ts file.

Having the guessed language written into the initial .ts file matched
QMake behavior, but it turns out that it's not actually necessary. If
the language attribute is missing, Qt Linguist displays a dialog on
opening the file where the source and target languages can be selected.
And the guessed language is pre-selected.

We now write a minimal initial .ts file on configure time. This file
- can be read by lrelease when building the *_lrelease targets
- can be updated by lupdate when building the *_lupdate targets
- is not deleted on clean

Fixes: QTBUG-107853
Pick-to: 6.5 6.6
Change-Id: I5f67a0cf0f02dbf0237a15adf493e9d7aac2485c
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>