lupdate: Fix missing context in functions with trailing return type
...in the non-clang C++ parser.

Out-of-line function definitions with "complex" trailing return type
threw lupdate off the rails: the context could not be extracted.

For example, the following tr call did not get picked up properly:

auto Foo::bar() -> C<int>
{
    tr("x");
    return {};
}

This happened, because the "prospective context" was cleared when
encountering unhandled tokens (the angle brackets). We now try to detect
the situation whether -> starts a trailing return type and omit the
clearing of the prospective context while in a trailing return type
declaration.

For clarity, Tok_Other has been renamed to Tok_AngleBracket, because
that's what it really represents.

The lexer was extended to recognize the 'decltype' keyword. The parser
skips all tokes in the parentheses that follow decltype. It's highly
unlikely to encounter a tr call there. That allows to handle situations
like

auto Foo::bar() -> decltype([]() { return 1; })
{
    tr("x");
    return {};
}

Pick-to: 6.5 6.6
Fixes: QTBUG-110949
Change-Id: I05a13891195dd76eb89e59d968d5834d40704704
Reviewed-by: Kai Köhne <kai.koehne@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>