qdoc: Fix incorrect meta content generated for \brief containing a link
The HTML generator erraneously used output from generateText()
directly as the value of a <meta> attribute, for the purpose
of adding a \brief statement as the meta content. If the \brief
contained a link or formatted text, the output was ill-formed.

To fix, call Text::toString() on the retrieved \brief, and
ensure it's HTML-compliant with a call to protectEnc().

Fixes: QTBUG-114151
Pick-to: 6.5 6.6
Change-Id: Ib88f0cd31b4f0d88509f084e4ecae81152f91c9b
Reviewed-by: Paul Wicking <paul.wicking@qt.io>