qdoc: Sort the output of \generatelist <groupname>
When generating an unordered list of group members with the \generatelist
command, the output was not ordered.

Use the same method of ordering as \annotatedlist, Node::nodeNameLessThan()
as the comparison function.

In DocBookGenerator::generateAnnotatedList(), implement sorting as well.
This function is used for both annotated and simple (itemized) lists, and
sorting brings the DocBook output more in line with HTML generator.

Pick-to: 6.5 6.6
Fixes: QTBUG-114956
Change-Id: I5cf967b0ec92b872dcc5e61cc5ed63ec295dd4a1
Reviewed-by: Paul Wicking <paul.wicking@qt.io>