CMake: Make qttools tests standalone projects
Add the boilerplate standalone test prelude to each test, so that they
can be opened with an IDE without the qt-cmake-standalone-test script,
but directly with qt-cmake or cmake.

Boilerplate was added using the following scripts:
https://git.qt.io/alcroito/cmake_refactor

Manual adjustments were made where the code was inserted in the wrong
location.

Task-number: QTBUG-93020
Change-Id: I1e3ff2b1130cc1e682a8c9242e52ba0eda667c6a
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>