QtDesigner: unbreak unity-build-batch-size 100000
Exclude TUs that cause problems in a build where all of QtDesigner's
.cpp files end up in a single unity_0_cxx.cxx. This should ensure that
the build will forthwith not fail because someone added a new .cpp
file in the "wrong" position.

Of course, this is just a snapshot, with my configuration: Clang 15,
Ubuntu 20.04, -developer-build, C++23, -sctp, libc++ (which means I
can't link most of the tools, because of problems similar to
QTBUG-115032).

Task-number: QTBUG-115448
Pick-to: 6.6 6.5
Change-Id: I2363194bc2c9e07d75d58be7386dba9fb0778a2e
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>