QDoc manual: Refer to correct minimal configuration
The QDoc manual contains documentation about the configuration of QDoc
documentation projects. A minimal configuration file is part of this.

As QDoc has evolved over time, so has the requirements for the
configuration file. However, QDoc's manual hasn't kept up with the
changes; the page describing a minimal `.qdocconf`-file is from before
QDoc moved from qtbase.git into qttools.git (October 2015). A minimal
configuration file, `examples/minimum.qdocconf`, is also part of QDoc's
documentation. This was turned into copypasta in 2018. However, the
outdated documentation page was still retained, such that the QDoc
manual links users to a minimal configuration that is either useless,
or usable.

This patch removes the outdated configuration file page, and updates the
link to this page such that the documentation will refer to the minimal
configuration file (as is done elsewhere). This reduces the maintenance
burden of the QDoc manual, as there's now only one minimal example. It
also ensures that users will find helpful information when perusing the
documentation for QDoc.

Furthermore, the minimal configuration file that is kept is modified to
ensure users can copy the contents into their own configuration and
successfully run QDoc to build the documentation project. This involves
the following changes:

- Add the missing `project` configuration setting, as QDoc will bail
  withouth it.
- Add a simple `includepaths` block, as this may be confusing for new
  users.
- Add a `warninglimit` block. Although strictly speaking not required
  in a minimal `.qdocconf`, it can be a useful tool when building a new
  documentation set.
- Reflow some text, make sure it's more in line with QDoc anno 2023.

Fixes: QTBUG-115602
Change-Id: Iabca9c538ae303dd4da589b46f11e50b321d4291
Reviewed-by: Andreas Eliasson <andreas.eliasson@qt.io>
(cherry picked from commit 746cfebb0e0ca2d6ae94c2d82dfc85bb1c228575)
Reviewed-by: Paul Wicking <paul.wicking@qt.io>