QDoc: Reorganize src/qdoc/qdoc
"src/qdoc/qdoc" contains the source code for the QDoc executable, along
with the source code for QDoc's manual.

Those files initially lived all together under "/src/qdoc", but recent
rearregement of the "qdoc" directory have moved them one level deeper,
to make space for related-but-contained sibling projects under the
"src/qdoc" directory.

Along the lines of those recent changes, and to make space for tests to be
moved under the directory, the "src/qdoc/qdoc" directory
is reoganized to follow a similar structure to other projects in its
parent.

All of qdoc source code is now moved under "src/qdoc/qdoc/src/qdoc".
The source code for QDoc's manual is left under the "doc" directory in
"src/qdoc/qdoc/", to give it more visibility and to better separate
between changes to the source-code level documentation and the
additional documentation that the project provides.

The "CMakeLists.txt" in the "src/qdoc/qdoc" directory was modified to
refer to the new paths where necessary.

The "CMakeLists.txt" in the "test/auto/qdoc/" directory, which specify
certain test targets that depend on the paths in the modified qdoc
directory, was modified to respect the new structure.

`#include` directives that were affected by the new structure have been
modified to respect the new paths.

Change-Id: I8b939889d7134863a1d1a09a96446c2076ee5e58
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>