QDoc: Move tst_QDoc under src
QDoc is currently restructuring its source code to simplify certain
expected future developments.

As part of this restructuring, all tests have been moving from the
"tests" directory to an appropriate place under the "src" directory.

`tst_QDoc` tests some internal components of QDoc that are currently
only exposed as part of the source code for the QDoc's binary.

It was previously positioned under the "test" directory and is now moved
under the "src" directory, as part of the restructuring.

A new directory, "tests" was created under "src/qdoc/qdoc", to contain
tests that are related to the QDoc binary itself, which lives under
"src/qdoc/qdoc".

"tests/auto/qdoc/qdoc", that contains the source code for `tst_QDoc`,
was moved under the newly created "tests" directory.

The main "CMakeLists.txt" file in "src/qdoc/qdoc" was modified to refer
to the "tests" directory when Qt is expected to configure tests.

A new "CMakeLists.txt" file was created under "src/qdoc/qdoc/tests" that
dispatch the configuration to the tests present in the directory.
This means that it currently dispatches only to the definition of
`tst_QDoc` in the "qdoc" directory, but is expected to be expanded upon
as tests are moved from the repository-level "tests" directory.

The "CMakeLists.txt" file for `tst_QDoc` was modified to support the new
positioning in the file tree.

Mainly, each path in it was modified to be prefixed by
`${CMAKE_CURRENT_LIST_DIR}` to ensure their correct resolution when
using CMake's `include`.

The "CMakeLists.txt" file in "tests/auto/qdoc" was modified to include
"src/qdoc/qdoc/tests/qdoc/CMakeLists.txt", to ensure that `tst_QDoc` is
able to run in CI.

A comment was added to the "CMakeLists.txt" file in "tests/auto/qdoc" to
explain why we are including certain "CMakeLists.txt" files how the code
that perform this operation work, as it may be otherwise inscrutable
without some knowledge of the Qt Project's CI and the general structure
of a submodule in the project.

Change-Id: Ifb04c893341e2d074178df0f7cf89be927b2b7f8
Reviewed-by: Paul Wicking <paul.wicking@qt.io>