Eradicate the last Q_FOREACHs and mark the module free of them
For an rvalue, store in a const lvalue. This is risk-less.

For a loop over addAction(), use addActions() instead. This is
risk-less.

For a loop over a subobject of an argument passed by const-ref, the
body clearly doesn't modify the container we iterate over (inspection
of sourceCode() implementation).

For a loop over a subobject of a local variable, the body clearly
doesn't modify the container we iterate over; use std::as_const().

Change-Id: Iccff25263bd1124da26a230055d4d28570f5800e
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>