QDoc: Move tst_config under src
QDoc is currently restructuring its source code to simplify certain
expected future developments.

As part of this restructuring, all tests have been moving from the
"tests" directory to an appropriate place under the "src" directory.

`tst_config` tests some internal components of QDoc that are currently
only exposed as part of the source code for the QDoc's binary.

It was previously positioned under the "test" directory and is now moved
under the "src/qdoc/test" directory, as part of the restructuring.

The "CMakeLists.txt" file under "src/qdoc/qdoc/test" was modified to
dispatch to the moved "config" test directory.

The "CMakeLists.txt" file for `tst_config` was modified to support the new
positioning in the file tree.

Mainly, each path in it was modified to be prefixed by
`${CMAKE_CURRENT_LIST_DIR}` to ensure their correct resolution when
using CMake's `include`.

The "CMakeLists.txt" file in "tests/auto/qdoc" was modified to include
"src/qdoc/qdoc/tests/config/CMakeLists.txt", to ensure that `tst_config` is
able to run in CI.

Change-Id: I771535b93e50ced315c6894c17feaed7a8fb662f
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>