QDoc: Move tst_qdoccommandlineparser under src
QDoc is currently restructuring its source code to simplify certain
expected future developments.

As part of this restructuring, all tests have been moving from the
"tests" directory to an appropriate place under the "src" directory.

`tst_qdoccommandlineparser` tests some internal components of QDoc that
are currently only exposed as part of the source code for the QDoc's
binary.

It was previously positioned under the "test" directory and is now moved
under the "src/qdoc/test" directory, as part of the restructuring.

The "CMakeLists.txt" file under "src/qdoc/qdoc/test" was modified to
dispatch to the moved "qdoccommandlineparser" test directory.

The "CMakeLists.txt" file for `tst_qdoccommandlineparser` was modified
to support the new positioning in the file tree.

Mainly, each path in it was modified to be prefixed by
`${CMAKE_CURRENT_LIST_DIR}` to ensure their correct resolution when
using CMake's `include`.

Special care was needed for the single resource file,
"tst_argument.txt", that `tst_qdoccommanlineparser` uses.
Specifically, the file is now specified through an absolute path, to
ensure that it correctly works through CMake's `include`, which in turn
require the `QT_RESOURCE_ALIAS` property to be set on the file, so that
the resource system can refer to it correctly.
Thus, the `QT_RESOURCE_ALIAS` was set for the relevant resource file.

The "CMakeLists.txt" file in "tests/auto/qdoc" was modified to include
"src/qdoc/qdoc/tests/qdoccommandlineparser/CMakeLists.txt", to ensure that
`tst_qdoccommandlineparser` is able to run in CI.

Change-Id: Ibea2bb83af753d2a2878407aecbc9782e23a9aea
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
Reviewed-by: Paul Wicking <paul.wicking@qt.io>