QDoc: Move tst_utilities under src
QDoc is currently restructuring its source code to simplify certain
expected future developments.

As part of this restructuring, all tests have been moving from the
"tests" directory to an appropriate place under the "src" directory.

`tst_utilities` tests some internal components of QDoc that are
currently only exposed as part of the source code for the QDoc's binary.

It was previously positioned under the "test" directory and is now moved
under the "src/qdoc/test" directory, as part of the restructuring.

The "CMakeLists.txt" file under "src/qdoc/qdoc/test" was modified to
dispatch to the moved "utilities" test directory.

The "CMakeLists.txt" file for `tst_utilities` was modified to support
the new positioning in the file tree.

Mainly, each path in it was modified to be prefixed by
`${CMAKE_CURRENT_LIST_DIR}` to ensure their correct resolution when
using CMake's `include`.

The "CMakeLists.txt" file in "tests/auto/qdoc" was modified to include
"src/qdoc/qdoc/tests/utilities/CMakeLists.txt", to ensure that
`tst_utilities` is able to run in CI.

Change-Id: Ia75bd1bcbb8669486ed7deef7a5f81eae20093cb
Reviewed-by: Paul Wicking <paul.wicking@qt.io>