QDoc: Move tst_generatedoutput under src
QDoc is currently restructuring its source code to simplify certain
expected future developments.

As part of this restructuring, all tests have been moving from the
"tests" directory to an appropriate place under the "src" directory.

`tst_generatedoutput` performs end-to-end regression testing for the
"qdoc" binary.

It was previously positioned under the "test" directory and is now moved
under the "src/qdoc/test" directory, as part of the restructuring.

The "CMakeLists.txt" file under "src/qdoc/qdoc/test" was modified to
dispatch to the moved "generatedoutput" test directory.

The "CMakeLists.txt" file in "tests/auto/qdoc" was modified to include
"src/qdoc/qdoc/tests/generatedoutput/CMakeLists.txt", to ensure that
`tst_generatedoutput` is able to run in CI.

Adds SPDX license headers to files that were missing it in "generatedoutput/testdata"

Change-Id: I412496ef0bf8ea9c07deda63960c11f0a5a06cc4
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>