CMake: Allow multiple calls of qt_add_lupdate
Do not use the target argument to construct the custom target's name,
because we want to specify multiple targets in a future commit. Instead,
use the project's name.

Support multiple calls of this command by appending an increasing number
to the default custom target name if the target already exists.

Make it possible to specify the custom target name in qt_add_lupdate and
qt_add_translations.

Fixes: QTBUG-115845
Change-Id: I20f08ab52eef3f430a4d9c20a0d900a2aec607c5
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>