CMake: Make the target argument of qt_add_lrelease optional
[ChangeLog][CMake] Passing a target as the first argument to
qt_add_lrelease was deprecated.

The target argument only served the purpose of driving the custom
command for lrelease. Instead, we use a custom ALL target for that
purpose. That custom target's name follows the same naming rules as the
one in qt_add_lupdate. Its name can be customized with the
LRELEASE_TARGET argument.

This is done to avoid having to pass a single target to qt_add_lrelease
in a future version of qt_add_translations that takes a list of targets.

There is Apple-related logic that extracts the languages from the .ts
files and stores them in a target property. This is still done, but only
for the "legacy signature" of the command that takes a target. The logic
was extracted into a function to be called in the multi-target version
of qt_add_translations in a future commit.

Change-Id: I6916c32dbc542adf2554b860f906fd94e3f4783a
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>