QDoc: Fix regression in DocParser::getRestOfLine()
DocParser::getRestOfLine() was refactored recently. The refactoring
patch is responsible for making the method unconditionally return a
`simplified()` QString, while this happened conditionally in the past.
This is interferes with QDoc's regular expression capabilities, as
QString::simplified() removes "padding" whitespace characters, often
used as indentation in code snippets, etc.

The refactoring patch integrated successfully due to the lack of tests
exercising QDoc's commands that take regular expressions as input.

This patch adds test data from the related bug report to the generated
output test for the affected QDoc command `\printuntil`. The test data
contains trailing whitespace, which is not expected to end up in output.
`DocParser::getRestOfLine` is modified to return a `trimmed()` instead
of `simplified()` QString, as this drops leading and trailing whitespace
only, and leaves other whitespace in the string untouched.

Done-with: Topi Reiniö <topi.reinio@qt.io>
Fixes: QTBUG-114937
Change-Id: If9b65b1a2d036a41057c572f1b42c5e61784efd5
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>