qdoc: Fix missing arguments in Node::signature() output
As part of a previous cleanup commit 4c3108e5c2, an overload of
Node::signature() was introduced. This overload took two arguments,
but it was not overridden by FunctionNode, causing the base
implementation to be called at multiple call sites. This base
implementation simply called Node::plainName(), causing a plain
function name to be listed instead of the full signature that
includes argument types and names.

Remove the problematic overload, and modify the remaining signature()
to accept a single argument; a new flag type that controls the parts
to include in the reconstructed signature. This makes the call site
code more readable, compared to three boolean arguments.

Add a test case for the \reentrant command as this problem (missing
arguments in the signature) was associated with its output.

Fixes: QTBUG-115116
Change-Id: I70352532297b25b4c38249c0a724050cad47b07f
(cherry picked from commit e47cb8bb01dd450d00937520022f25dd3d34dde0)
Reviewed-by: Paul Wicking <paul.wicking@qt.io>