CMake: Support project-wide i18n with qt_add_translations
Before this patch, qt_add_translations took one target argument. This
target was used for determining the sources that contain translatable
strings. The target was also considered as the only target that intends
to load .qm files. This "one target approach" is too limited.

Projects may have multiple executables that load .qm files. To allow
this, it is now possible to pass multiple targets with the TARGETS
argument. The old signature is still working for convenience.

Projects may have multiple targets that contain translatable strings.
Multiple such targets can now be specified with the SOURCE_TARGETS
argument. If SOURCE_TARGETS is not given, qt_collect_i18n_targets is
called to determine such targets.

To be able to automatically collect i18n targets that are defined after
the call to qt_add_translations, the command defers itself to the end of
PROJECT_SOURCE_DIR's scope. This, however, is only possible for CMake
3.19 and newer. With older CMake versions, the user will see a warning
with a hint how to call qt_add_translations non-deferred (IMMEDIATE_CALL
keyword).

[ChangeLog][CMake] qt_add_translations gained the TARGETS argument to
specify multiple targets that intend to load .qm files.

[ChangeLog][CMake] qt_add_translations gained the SOURCE_TARGETS
argument to specify multiple targets that contain sources with
translatable strings. A facility to automatically collect SOURCE_TARGETS
was added.

[ChangeLog][CMake] If the first argument of qt_add_translations is a
target, that target is not used anymore for extracting translatable
sources.

[ChangeLog][CMake] qt_add_translations now defers itself to the end of the
scope of PROJECT_SOURCE_DIR if SOURCE_TARGETS is not given. This
behavior can be turned off by passing IMMEDIATE_CALL.

Task-number: QTBUG-95183
Change-Id: If4419bb66eb4a6abe906e774a539c781612b73d1
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>