QDoc: Don't add deprecated set/is/has to see also
When generating documentation, QDoc can automatically add links to
certain non-macro functions to the "See also" section of documented
entities. Specifically, it may add links to the documentation of `foo`
for `setFoo`, `isFoo`, and `hasFoo` functions, if they exist. However,
this doesn't account for whether the function that's added to the see
also list is deprecated. Thus, QDoc may generate links to deprecated
content.

This patch modifies QDoc's `Generator::supplementAlsoList` method by
inserting a check for the deprecated status of the node QDoc is
currently generating documentation for as well as of the node considered
an additional see also target. In cases where none are deprecated, or
where both are deprecated, behavior is unmodified. In cases where the
node QDoc is generating documentation for isn't deprecated but the
candidate for linking to is, don't add the link.

This patch has minor implication for the documentation of Qt:
  qapplication.html     |    1 !
  qlistwidgetitem.html  |    1 !
  qtablewidgetitem.html |    1 !
  qtreewidgetitem.html  |    1 !
  4 files changed, 4 modifications(!)

These changes are expected from the behavior change that follows from
this patch.

Pick-to: 6.6
Fixes: QTBUG-116412
Change-Id: I5edbd33183746202cf058e7c80faf3a0c3dbf205
Reviewed-by: Kai Köhne <kai.koehne@qt.io>
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>