Move TR_EXCLUDE handling for source files from lprodump to lupdate
We want to use the same, wildcard-based i18n exclusion pattern in our
CMake translation API. Without this patch, we would have to do the
wildcard-based source file filtering in CMake and try hard to make it
equivalent to lprodump's filtering.

Instead, move the filtering to lupdate. Now, we can pass the full source
lists and exclusion patterns to lupdate using the JSON project file
format, and lupdate will take care of the source file filtering.

Change-Id: I9f73e057f9c29a179b4d7989577d3ea347e0c5e9
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>