CMake: Allow excluding target sources from i18n
[ChangeLog][CMake] The target property
QT_EXCLUDE_SOURCES_FROM_TRANSLATION was added to exclude source files of
a target from handling with lupdate.

The introduced target property takes a list of paths, absolute or
relative to the current source dir. Wildcards are supported.

The lupdate JSON format was extended with a 'sourcesExcludePatterns'
key.

Task-number: QTBUG-80848
Change-Id: I178a4277eed2b8837cfb68f91ddaff7169965410
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>