qdoc: Ensure a group is listed as part of the navigation bar
QDoc adds a link to a documented \group as part of the navigation bar
if a page uses an \ingroup command. However, this happened only if
there was exactly one \ingroup command used.

Pages can belong to multiple groups. Often, the \ingroup command is
used for grouping pages for use with \generatelist or \annotatedlist
commands, without any associated \group page.

Adjust the logic for resolving the group page link on the navigation
bar; select the first one that references a documented (and titled)
\group.

Document the automated link generation for \ingroup, and touch up
the related documentation.

Fixes: QTBUG-116441
Pick-to: 6.5 6.6
Change-Id: Ie773ac4f52733ca2a9ade04c96101e05ef05c0cb
Reviewed-by: Paul Wicking <paul.wicking@qt.io>