Fix race condition in test_translation_api CMake test
Let the '*_lrelease' targets depend on 'update_translations'. This
avoids the situation where lupdate has not finished writing a .ts file
and lrelease already attempts to read it.

This is necessary, because we want the test to automatically build
'update_translations' which modifies .ts files, which are considered
source files. User projects don't have this problem, because they will
build 'update_translations' separately.

Fixes: QTBUG-116599
Change-Id: Icdc88b0f4824efa8768770bd991af2ef2940fff1
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>