QDoc: Improve recognition of C++ default arguments
When QDoc parses the source code of a project, it retrieves and stores
the presence, or lack thereof, of default arguments for callables, so as
to be able use them for certain internal analysis and to later present
the information to the user.

It does so through libclang.

When examining the cursor for a callable, it traverses the cursor's tree
to find all parameter declarations.
When a parameter declaration is found, its tree is traversed to look for
an expression, which is expected to be the initializer for the default
argument.

While this approach generally works, it has a few degenerate cases.
The most prominent in Qt's documentation is related to fixed-size array
types.

When a parameter has a fixed-size array type, such as `int p[S]`, `S` is
an expression that will be found when traversing the parameter
declaration tree.
With the above approach it will be considered a default argument
initializer even when such is not the case, or when no default argument
is present.

In other cases, the approach fails to correctly handle the expression so
that it cannot retrieve the underlying code. In those cases, QDoc uses a
literal "...", that is later shown to the user, in place of the actual
initializer.

To avoid the above issues, the retrieval of default arguments was moved
to Clang's C++ API instead of libclang.

`get_cursor_declaration`, the entry point to lower a libclang `CXCursor`
to `clang::Decl`, was moved up in its file, so that it would be visible
to the new usages.

`ClangVisitor::readParameterNamesAndAttributes`, that takes care of
performing the above mentioned process, was modified to lower the found
parameter declaration cursor to a `clang::ParmVarDecl`, from which the
default argument is extracted using the common and more accurate methods
of Clang's C++ API.

Clang sometimes consider the `=` token in a default argument declaration
to be part of the source range of the argument initialization itself.
As QDoc expects this to not be the case, we manually remove the token
from a stringified version of the relevant source code.

Similarly, we remove the leading and trailing spaces from the same
stringified version to improve the presentation and to simplify the
later processing of the initializer.

The new method ensures that we don't incorrectly read a non initializer
expression as being a default argument initializer anymore.

Furthermore, it broadens the amount of initializers that we can
correctly find, removing the case, at least for Qt's current
documentation, where QDoc presents a literal "..." to the user.

Task-number: QTBUG-111970
Change-Id: Ibf6b2d91e94adefdbb419926a6071e787951584c
Reviewed-by: Paul Wicking <paul.wicking@qt.io>