qdoc: Fix incorrect fragment identifier appearing in resolved links
When resolving a link target for a string, QDoc does it in stages; first, it checks if a target matches a page title. Then, it matches
against a stored map of target records that include TOC entries (section
titles), \keyword, and \target entries. Finally, a path constructed
from the string (e.g. Module::Type::Member) is matched against an API
entity.

Of these, section titles (TargetType::Contents) are handled in a special
way: Instead of returning a match immediately, QDoc checks for an API
match and only falls back to the section title if an API match fails.
This is done to give a higher priority to API entities over section
titles as link targets.

QDoc stored the type of the matched item from the map and returned it
to the calling function. However, it failed to clear the target type
when it found a higher-priority target. This, in turn, lead to QDoc
using the incorrect ref (fragment identifier) for the final link
target, if the link string matched a section title earlier in the
process.

To fix, set the targetType in Tree::findNodeForTarget() only when
returning a matched Node instance.

Streamline the implementation by extracting a recurring piece of logic
as a lambda. Remove unused enum values from TargetRec::TargetType
- these are remnants that only cause confusion.

Pick-to: 6.6 6.5
Fixes: QTBUG-116335
Change-Id: I970c2b121858ad228ef21dead51a3edb30e86450
Reviewed-by: Luca Di Sera <luca.disera@qt.io>