QDoc: Improve recognition of anonymous enums
When QDoc parses the code for a C++ project, it tries to recognize
anonymous enums, so as to be able to treat them specially to allow the
user to document them.

To do so, when an enum declaration cursor is found by libclang, QDoc
checks that the name of the type of the enum is empty.

This works because the common way to get the name of a type,
`clang_getTypeSpelling`, would return an empty string for the type of an
anonymous enum.

While this generally worked during the years, the behavior of
`clang_getTypeSpelling` changes in LLVM 16, so that anonymous enums'
types are represented differently, which breaks the current code for
QDoc.

To avoid the issue, anonymous enums are now identified through the use
of `clang_Cursor_isAnonymous`, which is specifically made for this
purpose and is expected to keep compatibility between versions.

This changes nothing for QDoc on LLVM 15, keeping the exact same
behavior, while it aligns QDoc on LLVM 16 to that same behavior,
ensuring the presence of some forward compatibility.

Change-Id: I33c52c2da1c1e78caa2d2454aab7d9739fac9b09
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>