QDoc: Avoid marking noexcept(false) callable as noexcept
When QDoc parses the source code of a C++ project, it tries to extract
explicitly written noexcept information from relevant callables, so that
it can later show this information to the reader.

To do so, it uses Clang C++ API facilities to extract `noexcept(...)`
specifiers as written.

For certain cases in C++, in particular destructors, it might be needed
to specifically express that something is not `noexcept` by the usage of
a `noexcept(false)` expression.

Clang will consider those cases to have an exception specifier, which
QDoc will report to the user, producing a confusing `This function does
not throw any exception when "false" is true.` line in the output
documentation.

To avoid the issue, the relevant part of
`ClangCodeParser::processFunction`, which takes care of this process,
was modified to avoid marking a callable as `noexcept` if the `noexcept`
expression provided by the user is the literal `false`.

While a working solution, this might be a bit brittle in the presence of
some indirect always false condition, and might need some massaging in
if such a case appears in the future.

Fixes: QTBUG-116880
Change-Id: I3e11ffb73337fb67bd3a965b90e3c62119672cf6
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>