qdoc: Qualify proxy pages with the module name
When QDoc parses a \relates command, it tries to associate the topic
to another aggregate (class, namespace, or header). If this aggregate
doesn't exist, a proxy page is generated for the aggregate. These proxy
pages act as the output location for the documentation of related
non-members that live in a module different to where the actual
aggregate is documented.

The reference of the actual aggregate will then list the related
non-members, linking to the proxy page(s). However, the filename of the
proxy page was not qualified with the name of the module; if there
were more than one proxy page generated for a single aggregate
(because topics in multiple modules try to relate to it), then the
proxy pages end up being overwritten in online documentation where
QDoc uses a flat output directory structure.

Fixes: QTBUG-117510
Change-Id: I4a8b14fe7129ae885b5d2a440b51eee4ee35035f
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
Reviewed-by: Luca Di Sera <luca.disera@qt.io>
(cherry picked from commit 5f91518265e7f818be7f7c77ef1d18a24ca52bb8)
(cherry picked from commit bdb18ec1bde8b7cea2c09166b8d2c609602a2336)