CMake: Fix usage of qt_add_translations when targeting iOS
When targeting iOS, the usage of qt_add_translations led to the
following error:

  The custom command generating
    .../i18n_zh.qm
  is attached to multiple targets:
    i18n
    i18n_lrelease
  but none of these is a common dependency of the other(s).  This is not
  allowed by the Xcode "new build system".

This is the same as QTBUG-103470 which was already fixed, but a recent
change broke it (see below).

This amends commit 94bb328a17bc01d28ecb6464744cf7ef89c4055a.

We must not limit the fix for QTBUG-103470 to the "legacy signature" of
qt_add_lrelease.

Fixes: QTBUG-117406
Change-Id: Ia6fab2f9563e7add2e3111651a13183707b8a3eb
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>