QDoc: Stop encoding `*/` and `\..\.` in HtmlGenerator
In the late 2000s, Java bindings for Qt were released as the project
"qt-jambi". This was discontinued at the end of that decade and spun
off as an open source project, QtJambi, currently maintained by Omix
Visualization (https://github.com/OmixVisualization/qtjambi).

When generating HTML output, QDoc replaces certain characters that are
reserved in HMTL with their HTML entity counterparts. This is done in
the `HtmlGenerator::protect()` method. However, the method also contains
a special case that replaces certain characters that /aren't/ reserved
in HTML, with their HEX counterparts without leading `0`. The sparse
documentation for this special case suggests it is there to satisfy the
Javadoc generator.

The special case is intended to apply to the sequence `*/`, and to a
full stop (`.`) if two full stops occur with any one character between
them. Going by the code comment, the intention of the latter is to
escape the last full stop in the abbreviations of the latin expressions
`id est` (`i.e.`) and `exempli gratia` (`e.g.`).

The implementation is rather naive; for the `*/` sequence, the check
is buggy (off-by-one error in the check for the characters `*` and `/`).
For the spaced dot pair, the implementation fails to take into account
its impact on e.g. ellipses or `6.6.6`.

The existence of this special case for the Javadoc generator may suggest
that QDoc featured a Javadoc generator in the past, which is something
it certainly doesn't feature now. It may also suggest that output from
QDoc was used as input to Javadoc in some form or another. Since QDoc
with Qt 5.15, QDoc features a generator for DocBook, which is more
suitable for use as input to some other generation tool than HTML is.

This patch removes the special case branch. This may cause QDoc to
generate `.` instead of `&#x2e;` in its HTML output in multiple
locations. This is to be expected.

Change-Id: I45734c11ad5891016dd19b211d54f218c2def3a1
Reviewed-by: Luca Di Sera <luca.disera@qt.io>
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
(cherry picked from commit 2e06a8546ffbb8fb05817930553f83b35d0a0d55)
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit 660c7e29f7fafcee034073db2ebceececd20c352)