QDoc: Ensure section command doesn't shadow inherited module state
When generating documentation for a module that is in a state, for
example, "Technological Preview", as defined by QDoc's `\modulestate`
command, classes that belong to that module are supposed to feature
an entry in their requisites table (along with header, CMake/qmake
lines, etc).

If the documentation for such a class contains a QDoc command that
triggers the creation of `DocPrivateExtra`, such as any `\section`
command, the entry in the requisites table disappears. This is due to
the fact that `Generator::formatExtra()` categorically returns the
status that's recorded in the `metaTagMap` of the `DocPrivateExtra`, if
one exists, even if it's empty. As such, when e.g. a `\section` command
is encountered, an (empty) such `metaTagMap` comes into existence for
that documentation block.

This patch contains a minimal QDoc project that reproduces the
misbehavior, and that serves to confirm that the bugfix contained
herein solves the observed issue.

The behavior correction ensures that `Generator::formatExtra` checks
that the status obtained from `DocPrivateExtra::metaTagMap` isn't
empty. If it is, allow processing the other possibilities. Else, return
early.

Pick-to: 6.6
Fixes: QTBUG-117778
Change-Id: I47525f69e31d0a924ca9ab380c6c7255040eca6c
Reviewed-by: Kai Köhne <kai.koehne@qt.io>
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>